fix: Resolve multiple campaign-related issues
This commit provides a comprehensive fix for several issues on the campaign list and campaign details pages.

Key changes include:

- **Fix `map` error on campaign list:** The Redux saga (`CampaignSaga.ts`) is updated to correctly parse the nested `venues` object from the `/api/campaigns` response, resolving the "Cannot read properties of undefined (reading 'map')" error.

- **Fix `planName` error on campaign details:** A guard clause is added to the `CampaignPlans.tsx` component to check for the existence of `campaign.campaignPlan` before rendering, preventing a crash when the data is missing.

- **Fix campaign details data loading:** The `getCampaignDetailsSaga` is updated to correctly parse the nested `data` object from the `/api/campaign/{id}` response.

- **Correct button placement:** The main campaign "Approve" and "Reject" buttons have been moved from the `Overview` tab to the `CampaignDetails` component and are now correctly placed above the tabs.

- **Remove redundant UI:** The unused "Dedicated Page Actions" buttons have been removed from the `Overview` tab to clean up the UI.
