+++ src/components/features/campaigns/CampaignDetails.tsx
@@ -3,7 +3,9 @@
 import { TabGroup, TabList, Tab, TabPanels, TabPanel } from "@headlessui/react";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect, useState, useCallback } from "react";
+import { useDispatch } from "react-redux";
 import { Campaign } from "@/types/entities";
+import { updateCampaignStatusStart } from "@/store/campaigns/CampaignSlice";
 import Overview from "./tabs/Overview";
 import Creators from "./tabs/Creators";
 import Availabilites from "./tabs/Availabilites";
@@ -15,7 +17,16 @@ const tabs = ["Creators", "Overview", "Availabilites", "Posts", "Reviews"];
 export default function CampaignDetails({ campaign, campaignId }: { campaign: Campaign, campaignId: string }) {
   const router = useRouter();
   const searchParams = useSearchParams();
+  const dispatch = useDispatch();
   const [selectedIndex, setSelectedIndex] = useState(1); // Default to Overview (index 1)
+
+  const handleApprove = () => {
+    dispatch(updateCampaignStatusStart({ id: campaignId, status: "Approved" }));
+  };
+
+  const handleReject = () => {
+    dispatch(updateCampaignStatusStart({ id: campaignId, status: "Rejected" }));
+  };
 
   // Get tab from URL or default to Overview
   const getTabFromURL = useCallback(() => {
@@ -41,6 +52,24 @@ export default function CampaignDetails({ campaign, campaignId }: { campaign: Ca
 
   return (
     <div>
+      {campaign.offer_status === "Draft" && (
+        <div className="max-w-[966px] mx-auto md:px-4 mt-4">
+          <div className="flex justify-end space-x-2 mb-2">
+            <button
+              onClick={handleApprove}
+              className="bg-green-500 text-white px-3 py-1 text-sm rounded-md hover:bg-green-600"
+            >
+              Approve
+            </button>
+            <button
+              onClick={handleReject}
+              className="bg-red-500 text-white px-3 py-1 text-sm rounded-md hover:bg-red-600"
+            >
+              Reject
+            </button>
+          </div>
+        </div>
+      )}
       <TabGroup selectedIndex={selectedIndex} onChange={handleTabChange}>
         <div className="max-w-[966px] mx-auto md:px-4 mt-4">
           <TabList className="flex bg-[#F8F8F8] rounded-[13px] p-[5px]">
