+++ src/store/campaigns/CampaignSaga.ts
@@ -26,7 +26,7 @@ import {
 function* getCampaignsSaga(action: GetCampaignsAction) {
   try {
     const response = yield call(axiosInstance.post, '/api/campaigns', action.payload);
-    yield put(getCampaignsSuccess(response.data));
+    yield put(getCampaignsSuccess(response.data.venues));
   } catch (error: any) {
     yield put(getCampaignsFailure(error.message));
   }
@@ -35,7 +35,7 @@ function* getCampaignsSaga(action: GetCampaignsAction) {
 function* getMoreCampaignsSaga(action: GetCampaignsAction) {
   try {
     const response = yield call(axiosInstance.post, '/api/campaigns', action.payload);
-    yield put(getMoreCampaignsSuccess(response.data));
+    yield put(getMoreCampaignsSuccess(response.data.venues));
   } catch (error: any) {
     yield put(getCampaignsFailure(error.message));
   }
@@ -55,7 +55,7 @@ function* getCampaignDetailsSaga(action: GetCampaignDetailsAction) {
   try {
     const { id } = action.payload;
     const response = yield call(axiosInstance.get, `/api/campaign/${id}`);
-    yield put(getCampaignDetailsSuccess(response.data));
+    yield put(getCampaignDetailsSuccess(response.data.data));
   } catch (error: any) {
     yield put(getCampaignDetailsFailure(error.message));
   }
